[Refactor] 로직 및 뷰 리펙토링
