+++ BookBridge/BookBridge/Model/Enum/TapCategory.swift
@@ -1,0 +1,14 @@
+//
+//  TapCategory.swift
+//  BookBridge
+//
+//  Created by 노주영 on 2/6/24.
+//
+
+import Foundation
+
+enum TapCategory: String, CaseIterable {
+    case find = "구해요"
+    case change = "바꿔요"
+    case recommend = "추천도서"
+}
